fix: git 오류 수정
